renaming file and splitting into 2 tests
